move var strusted to vars_distributed.py, add optimizer's block name, test=develop
